Database tables are in final constant and can be replace by other name, Hungarian language, search result (googleResults) store in db, migrate program code TODO!!! but the field is now exists, if we are run as a new instance! And I added a notification url for nofity a (personal or other) website if a search is completed. (I need this for create more stuff when a search is finished so i thing it's a good feature for other people.

+++ core/src/main/java/com/serphacker/serposcope/models/google/GoogleRank.java
@@ -22,7 +22,7 @@ public class GoogleRank {
     public final short diff;
     public final String url;
     
-    public GoogleRank(int runId, int groupId, int googleTargetId, int googleSearchId, int rank, int previousRank, int hits, String url) {
+    public GoogleRank(int runId, int groupId, int googleTargetId, int googleSearchId, int rank, int previousRank, long hits, String url) {
         if(previousRank == 0){
             previousRank = GoogleRank.UNRANKED;
         }
